Separate different summation functions into their own modules
Don't force users to always import all implementations if they want any one implementation
