+++ src/functions/sumShew.js
@@ -1,71 +1,5 @@
 import {expecting, wrap} from "../core/wrap";
 
-// Perform a linear summation of values in a sequence.
-// Suitable for integers and low-accuracy floating point sums.
-// Returns 0 for empty inputs.
-// For users who really can't be bothered to know how the various
-// summation functions are differentiated, stick to the principle of least
-// astonishment. Most users are acquainted with linear summation, perhaps
-// less so with Kahan and Shewchuk algorithms: alias "sum" to "sumLinear".
-export const sumLinear = wrap({
-    names: ["sumLinear", "sum"],
-    attachSequence: true,
-    async: true,
-    arguments: {
-        one: expecting.iterable
-    },
-    implementation: (source) => {
-        let sum = 0;
-        for(const value of source) sum += value;
-        return sum;
-    },
-});
-
-export const sum = sumLinear;
-
-// Compute a sum of numbers using the Kahan summation algorithm.
-// https://en.wikipedia.org/wiki/Kahan_summation_algorithm
-// More accurate than linear summation, but slower.
-// Less accurate than Shewchuk's summation algorithm, but faster.
-// If any input is NaN, returns the first NaN input.
-// If any input is +inf and no inputs are -inf, returns +inf.
-// If any input is -inf and no inputs are +inf, returns -inf.
-// If any input is +inf and any input is -inf, returns NaN.
-// In case of intermediate positive overflow, returns +inf.
-// In case of intermediate negative overflow, returns -inf.
-export const sumKahan = wrap({
-    name: "sumKahan",
-    attachSequence: true,
-    async: true,
-    arguments: {
-        one: expecting.iterable
-    },
-    implementation: (source) => {
-        let sum = 0;
-        let compensation = 0; // Error compensation term
-        let overflow = false; // Track intermediate overflow
-        for(const value of source){
-            if(isNaN(value)){
-                return value;
-            }else if(!isFinite(value)){
-                if(overflow || isFinite(sum)){
-                    sum = value;
-                }else{
-                    const n = value + sum;
-                    if(isNaN(n)) return n;
-                }
-            }else if(isFinite(sum)){
-                const y = value - compensation;
-                const t = sum + y;
-                compensation = (t - sum) - y;
-                sum = t;
-                overflow = !isFinite(sum);
-            }
-        }
-        return sum;
-    },
-});
-
 // Compute a sum of numbers using Shewchuk's summation algorithm.
 // http://stackoverflow.com/a/2704565/3478907
 // http://code.activestate.com/recipes/393090-binary-floating-point-summation-accurate-to-full-p/
@@ -149,4 +83,4 @@ export const sumShew = wrap({
     },
 });
 
-export default sumLinear;
+export default sumShew;
